+++ src/lexer.c
@@ -636,7 +636,7 @@ static tmbchar LastChar( tmbstr str )
 {
     if ( str && *str )
     {
-        int n = TY_(tmbstrlen)(str);
+        size_t n = TY_(tmbstrlen)(str);
         return str[n-1];
     }
     return 0;
